Completed implementation of merge sort and pivot sort
Compelted my implementation of merge sort and pivot sort algorithms will be working on radix sort next. Also did learn something interesting today. It seems that Java does have some pass by reference features in that for objects we can manipulate them to a degree as long as we aren't setting them to a new memory address or aren't manipulated primative data types. Discovered this was possible while coding up pivot sort because I had assumed that everything in java was pass by value. Well everything in java is pass by value but there is a limited scope for pass by reference in terms of objects.