Started the implementation of a singly linked list
Started the implementation of a singly linked list. Completed so the ability to add values at the end, start, and at specific index in a linked list. Also completed the ablity to return the value of the head and remove it from the linked list. Currently in the process of coding up the remove at a specific index functionality in the linked list.