Completed coding of Singly Linked List
Completed coding of singly linked list. Impleted the ability to remove values the beginning, end, or at a specific point. Also added in the ability to get the value fo the first node, last node, or at a specific index. Another thing added in was the ability to alter the value of a specific node. Finally added in the ability to reverse an entire linked list