Fixed the final bug with binary heaps
There was one more bug found with the binary heap when the values come down to just the final 3 the program starts freaking out and not switching values so for the min heap it could leave a higher value on the root and for a max heap a lower value on the root. Resolved the issue by altering a conditional operator that checks to see if the index is greater than the length of the min or max heap. Altered the value to check to see if the index is greater than or equal to the length of the min or max heap. The resolved the issue and allowed swapping to occur and function properly.